[REM] *: accounting files not in their module's manifest
*: account_sepa_direct_debit, l10n_ec, l10n_nl_reports

Description of the issue this commit addresses:

[This PR](odoo#211150) has brought our attention to some files that were in the codebase but
not included in their module's manifest. Therefore they are useless as is and
can either be deleted or need to be put in the manifest.

Desired behavior after this commit is merged:

Unused useless files have been removed from the codebase.
Unused useful files have been added to their module's manifest.

task-4822341

closes odoo/enterprise#86695

X-original-commit: 36db509de497a880da8bf6e440b9e19fe7451bde
Related: odoo/odoo#212348
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>