[IMP] account_accountant: allow user to click on the line to select
When you are in a list view, you can select multiple records by clicking the checkbox. But if after selecting multiple records, you click on a line, it will only select the record you just clicked and all other preselected records will be ignored.

This commit change that, as now if you already have selected records, if you clik on a line without checking the checkbox, it will add the record to the selection just like you clicked on the checkbox.

task-4816580

closes odoo/enterprise#86923

X-original-commit: 89e32626bbd5361444119d47604704a53e3e5f6d
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>