[IMP] l10n_us_reports: add US tax report variant
Before this PR, the Generic Tax Report had always been considered as
the only report necessary to handle the US taxes. The problem is that,
if a company needs to submit tax reports in multiple countries
(because it has a VAT number in them; what we call multivat in the
Odoo jargon), the Generic Tax Report would then show the taxes from
every company, not only the US ones, hence becoming a bit messy. This
also causes issues with the tax returns, which need to know the
country they belong to in order to properly set the tax lock date.

We solve that by introducing a new US tax report. It does the same as
the Generic Tax Report, but shows only the US taxes.

closes odoo/enterprise#87346

Related: odoo/odoo#213649
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>