[IMP] documents{_hr|_spreadsheet}: improve share panel
We convert the share panel in a wizard and improve it by allowing sharing
multiple documents at once.

To simplify the wizard, we allow it to be opened only if the user has edit
permission on all documents. Indeed, without edit permission, the user can only
copy the documents URLs and there is already a dedicated action for that.

Technical notes:
- if we set partner_id and documents_share_id fields as required in
documents_share_access, it doesn't work (ex.: wizard button not updated when
deleting an access).
- if we set original_role as required in documents_share_access, it causes a
validation error when updating the role when the wizard is saved (original_role
is empty while it has been initialized with a value). We cannot set it readonly
either (cause "format requires a mapping" error).
- we use a warning notification when inviting partners without email because we
cannot hide the "Share" button when selecting a partner without email (using
a computed field to disable it doesn't work probably because the pop-up to
complete the missing email is opened).

Task-4848234

Part-of: odoo/enterprise#87971
Related: odoo/upgrade#7959
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>