[IMP] quality_control: use form view for on demand quality check
This commit apply the following:
1- updates the on-demand quality check to use the form view, not the wizard.
2- applies pre-filling the necessary fields in the quality check form,
such as the Picking, Production Order, Product and Quality Team.
3- shows quality check Title field and makes it editable in the form view.
4- updates measure_on field to be the same as the quality point.
5- shows quality checks as todo even if they include products that
are not in the picking.
6- hides some fields in certain contexts.

NB. It is expected to have only one of the following fields set:
[picking_id, production_id, repair_id]. The user is warned about this
in case the DB had faulty records and is given a chance to fix it by
unsetting them to have none or only one of them set.

Upgrade PR: https://github.com/odoo/upgrade/pull/7774

Task-4535415

Part-of: odoo/enterprise#86685
Related: odoo/upgrade#7774
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>