[IMP] knowledge: allow redirection for trashed articles
Currently, when a user sends an article to the trash, the system triggers
a notification email to relevant members informing them that the article
has been trashed and will be permanently deleted in 30 days (see:
`knowledge_article_trash_notification`).

This email includes a link to the trashed article (see: `article_url`).
However, when a user clicks the link, they are taken to a 403 error page
and cannot access the article to review or restore it.

Steps to reproduce the issue:
1. Open an article
2. Send the article to the trash
3. Open the notification email and click on the article link
=> The user is redirected to a 403 error page

When an article is trashed (see: `action_send_to_trash`), it is marked
as inactive. As a result, it is excluded from ORM search results by default.

The `/knowledge/article/<id>` route uses the ORM to retrieve the article
by its ID. Since trashed articles are inactive, the ORM does not return
the article, leading the route to assume the article does not exist and
redirect the user to a 403 error page.

To resolve this issue, we will modify the context of the search method
to include the key `active_test=False`. This will allow the ORM to return
inactive (trashed) articles. As a result, users who follow the link from
the email will be correctly redirected to the trashed article page, where
they can view its content and choose to restore it using the banner interface.

Task-4636494

closes odoo/enterprise#86946

X-original-commit: 05500442d3a9f63bffad9ca34e821a77518e58f1
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>