[IMP] documents: implement dynamic traceback folder
The "Support" folder for tracebacks should be created dynamically
when (if) needed, and still be unavailable to portal users.

This commit includes a Python controller test for the route responsible
for uploading traceback files into the "Support" folder and a JS test
for the availability of the Share button.

Task-4690653

closes odoo/enterprise#82891

Related: odoo/upgrade#7846
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>