[IMP] l10n_pe_edi: support line and global discounts in export
Line level discounts were skipped in the previous rewrite of l10n_pe_edi
as it had weird edge-cases at the time, so before we just manually
reduced the prices of the products by the discount to account for the
final cost of the invoice.

However, this was not technically allowed or compliant with SUNAT, as
such, this PR allows for the export of Global and Line Level discounts
properly.

Note that we had the field l10n_pe_edi_allowance_charge_reason_code
defined previously in this module but it was not used anywhere. It could
be used here for discounts, however, discounts will always be a specific
value depending if it's global or line level so there is no need to
allow the customer to select it or something different.

As there is a chance that the future withholding tasks will need this field,
it is staying but made hidden as it doesn't do anything for now.

task-4605358

closes odoo/enterprise#85914

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>