[IMP] account_reports: Show report header without waiting for the report
When opening any accounting report, we are waiting for the report to
load before showing its header.

This makes it so we are displaying the header when we have the
information for it and then displaying the report once we receive its
information.

We moved some values from the get_report_information to get_options
since we need those values in the header and without them, the header
will refresh and we will lose the modification done while the report
was loading.

task-4826507

closes odoo/enterprise#86716

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Co-authored-by: oco-odoo <oco@odoo.com>