[FIX] approvals: prevent edit approved approval
Issue:
------
An approval can be edited by the request approval owner
even if it has been approved.

Solution:
---------
Prevent the owner from writing on the approval
after validation except to cancel it.
Prevent owner to approve his own approvals.

Note:
An `ir.rule` is not possible because the owner
must always be able to write on his approval
(but in a controlled way, i.e. only for cancel).

task-4840746

closes odoo/enterprise#86775

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>