[FIX] l10n_in_hr_payroll: comment invalid demo data
The `all_l10n` standalone test started to fail on 2025-06-01 when
installing `l10n_in_hr_payroll`. A first attempt was made to blacklist
the module in the test itself. This attemps failed as the module is
auto_install on `hr_payroll` and most of `l10n_hr*` modules depend on
`hr_payroll`.

closes odoo/enterprise#86788

X-original-commit: 6206199145d73cd4d57de9ae118e95ef3c58b188
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>