[FIX] account_accountant: Bank Reco widget, fix button layout
This commit fix the button layout on mobile, in the bank reco widget. Before this commit, we only showed one button when the user unfold a line.

Now, the idea is, as long the line is not reconciled, to always show the 3 main buttons, "Set Partner", "Set Account" and "Reconcile".

The buttons "Batches", which is only available in mobile, is now in the dropdown menu.

task-4804584

X-original-commit: 0313f41f9d8b9f4a85d08bffe40beb3676873a84
Part-of: odoo/enterprise#87209
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>