[FIX] ai: fix the button alignment in AI chats
This commit fixes the alignment of some buttons in the AI chats.
Namely, it moves the copy button for user messages from the left side of
the chat to the right under the user message. It also moves the AI action
buttons to the left to be more aligned with the start of the text bubble.

--- Behaviour Before this Fix ---
The copy button would be aligned on the left side of the chat window
while the user message would be on the right side. The AI actions buttons
would be slightly more to the right than the text bubble.

--- Changes Done for the Fix ---
For the copy button: Previously, the flex-reversed class was added in the
contentContainer div, which includes the message and its actions. That
was removed and now the flex-row-reversed class is added when the chat
is not AI. When the chat is AI, the flex-row-reversed class is added but
this time only for the actions div.
For the AI action buttons: instead of adding margins horizontally for the
buttons, we are only adding them to the right of the buttons. Also for
the actions div, we remove the left margin. We also added some padding
to the buttons' text because it was a little crammed.

--- Behaviour After the Fix ---
The copy button is now aligned on the right side, under the user's
message.
The action buttons are better aligned with the message bubble

Task-4840542

closes odoo/enterprise#87638

X-original-commit: 88c1e75df78b80ac8ad3dd544d5790af7b722b72
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>