[FIX] partner_commission,sale_subscription_partnership: fix partnership
Corrects the behaviour of the member level assignment in membership when childs partners are concerned.
A partner buying a membership should forward the related commission plan to its children.
When removed, children should not keep the membership if the parent lose it.
This commit does not apply this when the membership is changed from the form view instead of the related sale order.

TASK-4801759

closes odoo/enterprise#90787

X-original-commit: db59451a14905594141e01b836784cd88a5ae8be
Related: odoo/odoo#220130
Signed-off-by: Vallaeys Valentin (vava) <vava@odoo.com>