[FIX] account_reports: ec sales list warning for duplicate VATs
Fixed the computation of `is_vat_duplicated` to avoid false positive in case of partners having both invoices and refunds in the selected period.

The current query returns several rows for the same partner in case it has invoices and refunds, or different tax code... So we can't rely on that query to compute is_vat_duplicated, and should instead post-process it.

closes odoo/enterprise#87382

X-original-commit: 51c3f5331593a1a7e5603d96f400269d5a3112b0
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>