[FIX] pos_blackbox_be: Handle multiple taxes on work products
When multiple taxes are configured on the "WORK IN" and "WORK OUT"
products (a common scenario in a multi-company setup), the check
for the tax amount would fail. The previous code attempted to access
the `amount` field directly on a multi-record tax recordset, which
results in an error.

This caused the Point of Sale session to fail on opening.

opw-4744514

closes odoo/enterprise#86912

X-original-commit: 3dbb6ea62fac0cff27c50e0ce11122b261cb0f61
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>