[FIX] delivery_iot: select printer for shipping labels
Printing shipping labels is performed from the backend, once the shipping info are received in the chatter.
The printing command is sent to the frontend via the user bus, then though longpolling to the iot box.

This commit adds the possibility to select a printer instead of choosing automatically the first (with the right
report associated) on the list. As the change is made on a stable version, we are using system parameters to
store the selected printer without adding a new field.

We associate a printer with the picking type, in order for to be able to have different printers by default
on different picking types.

closes odoo/enterprise#97174

Task: 4792491
X-original-commit: 59f9a215788e5728fa5f9f918b55f12a6cda055a
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>