[FIX] account_accountant: prevent error when no reconciliation models are found
Currently, an error occurs when all reconciliation models are deleted and the
user attempts to create a Transactions(bank statement lines).

**Steps to produce:**
- Install the `accountant` module.
- Navigate to `Accounting > Bank > Models (under Reconciliation)` and delete
  all reconciliation models.
- Go to bank transactions, Create a new `bank statement line` and
  attempt to `save` it.

**Error:**
`SyntaxError: syntax error at or near ')'  LINE 39:    AND reco_model.id IN ()`

**Root Cause:**
At [1], an `SQL query`  is executed that references
`Reconciliation models(reco_models)`, as `reco_models` is empty, this is causing
an error.

[1]
https://github.com/odoo/enterprise/blob/8289993c7d52a6cf71982aace87bcb11fde16967/account_accountant/models/account_bank_statement.py#L220

This commit ensures and prevent error when user add transactions without reconciliation model.

sentry - 6650696331

closes odoo/enterprise#88557

X-original-commit: 7ffc620d17e86c47d6d53ea70fe65fed3693c8a0
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Utsav Maru (utma) <utma@odoo.com>