[FIX] l10n_se_sie4_export: handle account changes on tests
From saas-18.2, some account changes on Swedish CoA when other module(s)
are installed have been breaking the test for the export (that exports
all the available account).

This commit adapts the expected test file to ignore this changes and
ensures the result always stays the same, no matter what modules are
installed on the database.

This fixes the currently recurring runbot errors from saas-18.2 and up:

[runbot-159821](https://runbot.odoo.com/odoo/error/159821)

closes odoo/enterprise#88252

X-original-commit: dc2d4bb4bcb41a675fce1a61e8bb8413ea95e767
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>