[FIX] iap_extract: extract banner not shown
If one uploads a vendor bill from the list view using the
"Upload" button, a banner is then shown on the form view
of the created move to reload the data that were extracted
via IAP. However, this banner is not shown if one creates an
empty move, then uploads via the attachments.

This is because when we arrive in the `onWillStart` of the
`StatusHeader`, we subscribe to a channel using the
`extract_document_uuid` which does not exist upon creating
a new empty move. Then, when we upload our document, the
move record is saved, and is sent to the backend. At this
point, the move exists, and a `extract_document_uuid` is
set. Then we get the OCR results (in `_check_ocr_status`)
and we have a new status, we send this status on the bus
but on a channel named `f"extract.mixin.status#{self.extract_document_uuid}"`
where the `extract_document_uuid` is set. However, in the JS part,
it was empty because it was created in the `onWillStart`.

To resolve this, we subscribe to a new channel to detect newly
created documents. We send the new `extract_document_uuid`
in this channel from the backend to the frontend. The JS
can now subscribe to the correct channel using an existing
`extract_document_uuid`.

Now, the JS correctly receives the updates of the Python
via the bus, and can update the view accordingly.

opw-4566944

closes odoo/enterprise#86961

X-original-commit: 2d080cbd611227466a6bc073e6c2acfe4c6db7d9
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>