[FIX] account_accountant: multi selection of invoice and residual amount
This commit will fix three different bugs:
- When selecting multiple invoice in the select create dialog view and that the
amount of the selected invoice was bigger than the open_balance. a partial was
created directly which shouldn't be the case

- When matching multiple invoice and then adding one more after, the open
balance was wrongly computed since we were using the amount residual to
calculate it and since the invoice is reconciled the amount residual was 0.
Due to this commit: https://github.com/odoo/enterprise/commit/8bc9b75dd9a44ee87eb3b222cbec82a9697d8e6c

- When having a statement line of a 100, adding 2 lines 75 and 25, when adding
the last one, the residual is 0 and so the amount was display twice because
sourceBalanceBiggerThanLineBalance was returning true.

no task id

Part-of: odoo/enterprise#87350
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>