[PERF] documents: Don't recompute `is_multipage` on copy
Description
-----------
When copying a document, a new attachment is created and linked to the
copied document. Since `documents.document.is_multipage` depends on
`documents.document.datas`, which is related to `attachment_id.datas`,
and `attachment_id` was just written to, `_compute_is_multipage` will be
recomputed during flushing before commit, as it's a compute stored
field.

This is problematic when copying multiple large documents, since this
compute reads the entire binary data to determine if it's more than one
page.

This commits explictly writes on `is_multipage` to implicitly add it to
the protected fields, which will prevent its' recomputation when
flushing.

Benchmark
---------
Copying 5 documents (PDF) of 50MB each takes in total:

| Before | After  | Diff   |
|--------|--------|--------|
| 1.61GB | 1.17GB | -27.3% |

Reference
---------
opw-4829372

closes odoo/enterprise#87364

X-original-commit: 7099ef22b5893aceb4311a77c9bf4e5b19f07b34
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>