[FIX] web_studio: multiple occurences of same field in subview list
Have a sublist in form with multiple occurences of the same field (the fist one is invisible, the second is visible)
In studio, edit that list, click on the field (it should be second one in the view when only visible elements are displayed)

Edit the label (or any attribute on that field)

Before this commit, the first field was actually edited.
This was because in that setting, the python-side edition computed the xpath coordinates in function
of the node's attribute, omitting its relative position.

After this commit, the right field is edited in the view.

opw-4772870
opw-4664501

closes odoo/enterprise#88001

X-original-commit: 45a045376f1a67bcaf9ce97e3f8da7e85deb339f
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>