[FIX] web_studio: prevent error when no approval rules exist
An error occurs when a user attempts to mark a scheduled activity as done after
the associated approval rule has been deleted.

**Steps to Reproduce:**
1) Install Sales and Web Studio modules.
2) Log in as Admin and use Studio to add an approval rule to the Sale Order’s Preview button.
>- Set Allowed Group to Access Rights.
>- Set Responsible User to Mitchell Admin.
3) In the Incognito Tab, login as Demo, open the same sale order and click on
   preview to create activity in chatter.
4) Delete the Approval Rule in the original tab.
5) Switch back to Demo and click Mark Done under Planned Activity in chatter.

**Error:**
`SyntaxError: syntax error at or near ')'`
 `LINE 1: SELECT id FROM studio_approval_rule WHERE id IN () FOR UPDAT. ^`

**Root Cause:**
The error occurs because the SQL query at [1] includes an empty tuple of rule
IDs `(all_rule_ids)`. An empty IN () clause in SQL results in a syntax error.

[1]- https://github.com/odoo/enterprise/blob/7ea45724e7689a0df11d20ace9c562788f5d19e3/web_studio/models/studio_approval.py#L366

**Solution:**
This commit avoids the error by ensuring that the SQL query only runs when
`all_rule_ids` is not empty.

sentry-6306636466

closes odoo/enterprise#93242

X-original-commit: 20b45f6c65c78a572a3f26b78f6ed458accf7c9f
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Het Patel (path) <path@odoo.com>