[FIX] l10n_mx_edi: avoid overriding the payment policy of an invoice with the policy of the partner.
**PROBLEM**
When confirming an invoice, the payment policy of the invoice was overridden by the payment policy assigned to the partner, and the user had to reselect the correct policy.

**STEP TO REPRODUCE**
1. On a fresh database, install the l10n_mx_edi module.
2. Select the "ZAPATERIA URTADO ÑERI" as your company (demo mexican company).
3. Set a payment policy on a client.
4. Create an invoice with this client, and set the payment policy equal to a different policy than the client.
5. Confirm the invoice, the payment policy field will take the value of the client policy, overriding what we set in the invoice.

**CAUSE**
https://github.com/odoo/enterprise/blob/3f42e4322dbd684d9431a172c11a369090b21c0c/l10n_mx_edi/models/account_move.py#L617-L619

The payment policy compute method is incorrect, the fallback order for the payment policy is wrong.

**FIX**
The order should be :
If the invoice is a refund, payment policy should be PUE.
Else, if the user set a payment policy in the invoice form, we should keep it.
Finally, if the partner has a payment policy, select it.

opw-4820139

closes odoo/enterprise#87103

X-original-commit: f31e0ff6e5e431a0f992e4fd07c604be4468be28
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Léo Gizard (legi) <legi@odoo.com>