[FIX] appointment: Correct upcoming meetings count
Issue:
When we create more than one event with different status for the same appointment type,
the allowed events container is only updated with the most recently created event.

The Purpose of this PR:
To update the count of event ids for the same appointment type.

Steps to reproduce on runbot:
create multiple events for the same appointment type. shuffle the events around
return to the appointments menu and see that the upcoming meetings count does not match

opw-4831384

closes odoo/enterprise#87036

X-original-commit: 8466260ca6d1a820fd7c864e3b470073363109bb
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Ian Dao (iada) <iada@odoo.com>