[FIX] voip: fix keypad stuck in `in call view`.
In VoIP, the keypad state wasn't updated after the user hangs up with
the other part. So, when making a new call, the keypad appears again
with its last state in the previous call. This commit fixes this by
reseting the keypad state everytime the user makes a call.

Task-4848365

closes odoo/enterprise#87453

X-original-commit: bf77c8c3c7c1003a0755fbd67c6c41fe2e9a91ff
Signed-off-by: Yuchen Huang (yhu) <yhu@odoo.com>