[FIX] account_reports: convert date_from and date_to from date to string
Convert the fields date_from and date_to from a datetime.date to a
string since they are not JSON-serializeable. This problem would
usually not happend since the ORM convert them to string when sending
them to the client. However when the server directly uses them, they
are not converted and are still in datetime.date which shows the
error.

task-3603619

closes odoo/enterprise#87380

X-original-commit: f5234c27498b04967243a82c5b4411e3a7ce97fd
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Léo Leclerc (leol) <leol@odoo.com>