[FIX] account_accountant: select create dialog multi currency
Before this commit, when using multi currencies the remaining amount was wrongly
computed.
Multiple cases are possible either:
- Journal is the currency of the company then it was working fine.
- Journal is in another currency for example YEN. With this commit if you had
multiple move lines with different currencies we will not compute the remaining
amount.
- Journal is in another currency (YEN) but the statement line has a foreign
currency could be the company currencies or another. If the foreign currencies
is the one from the company then it's working fine. If not, we will hide the
remaining amount if there is multiple currencies.

no task-id

closes odoo/enterprise#87350

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>