[FIX] l10n_it_reports: ensure payment journal entries are balanced with bank account entry
The Libro Giornale report was missing the bank (liquidity) line for payment entries, which caused the report to be unbalanced.

This happened because the report is based on the standard journal report, which skips liquidity lines by default.

To fix this, the `_get_export_lines_for_journal` method was updated to treat the bank line like a normal journal line, so it now appears in the report.

This ensures that payment entries show both the credit and debit sides as required.

task-4830113

closes odoo/enterprise#87689

X-original-commit: 4c3a739edbf5ca6c13185d21db35b7df9e40eb39
Signed-off-by: John Laterre (jol) <jol@odoo.com>