[FIX] account_online_synchronization: prevent error when Starting Date is empty
Currently, an error occurs when finding duplicate transactions in the bank.

Steps to Reproduce:

 - Install the accountant module.
 - In Accounting Dashboard, click on Bank.
 - In Actions, click Find Duplicate Transactions.
 - Clear the Starting Date field and click save.

TypeError: unsupported operand type(s) for -: 'bool' and 'relativedelta'

This error occurs when the system attempts to find duplicate transactions
without a Starting Date. When finding duplicate transactions, it tries to
calculate the prior_date using date_from - relativedelta(months=3) [1].
However, since date_from is False, this raises an error.

[1] https://github.com/odoo/enterprise/blob/acdba5ecf89ee8920f82e07e16fac2a0f9659b12/account_online_synchronization/models/account_journal.py#L314

This commit ensures that if the Starting Date is empty when finding
duplicate transactions, the system uses today’s date instead.

sentry-6653835511

closes odoo/enterprise#88095

X-original-commit: 0725d303c0fab9ef047d2d1c6bf77bab7f579a72
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>