[FIX] account_iso20022: Handle payment types S and X for Swiss ISO20022 payments
Following page 35 of the documentation here https://www.six-group.com/dam/download/banking-services/standardization/sps/ig-credit-transfer-sps2024-2.1.1-en.pdf

In Switzerland, SEPA and ISO20022 are considered as subsets of the Swiss standard, and it is hence possible to include within the Swiss file nodes that are generated following these formats. We didn't support that before this commit.

To allow that, we also remove the ability to force the value of the ChrgBr node when using SEPA ; other values as SLEV are not supported by that format anyway.

opw-4535542

closes odoo/enterprise#87658

X-original-commit: 7e3223fdd7b4d5184e7f7cb1e8bcebbdf94f802e
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>