[IMP] planning: make Planning Analysis Graph and Pivot views clickable
This commit enhances interactivity in the Planning Analysis report by enabling
users to click on graph bars or pivot cells and be redirected to the respected
planning.slot list view.

The report itself is based on the planning.analysis.report model, but the goal
is to redirect to the planning.slot list view on click. Simply removing the
disable_linking from the xml view was insufficient for this behavior.

To address this, we extended the renderer to override the openView method and
explicitly set the resModel to planning.slot when a click action is triggered.

task-4798544

closes odoo/enterprise#87157

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>