[FIX] account_reports: Create Tax Reports for fiscal position when creating taxes
When we create a fiscal position, we are checking if the user has the
taxes for that country create and display a button to have the user
download them.

Previously, we were trying to create the tax reports for that fiscal
position when creating it, however since the l10n_report is loaded
with the taxes, the creation of the tax return wouldn't be triggered.

Now, we are creating the tax returns after the user has downloaded the
taxes needed for that fiscal position.

task-4840412

closes odoo/enterprise#87903

X-original-commit: 18b5adf2e3d0f9f83b7f98c3c4b5b98d33d8dfdf
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Léo Leclerc (leol) <leol@odoo.com>