[IMP] web_gantt: add maintain buffer resheduling
In this commit:
- Critical path is removed
- right and left arrow to reschedule tasks forward and backward is removed and
replaced by:
- Manual Reschdule -> ignore the amount of time between dependent tasks, even if
there's a conflict
- Auto-Reschedule (Use Buffer) -> consume the amount of time between dependent tasks
unless there's a conflict
- Auto-Reschedule (Maintain Buffer) -> maintain the amount of time between dependant
tasks

For auto-reschedule (maintain buffer):
- The buffer corresponds to the number of working hours between tasks A and B according
to the working calendar of the project's company (else the current company). If the working
calendar is configured as flexible, refer to the real time between tasks A and B instead.
- Dependent tasks should be rescheduled so that the buffer is maintained, while taking into
consideration the assignees' availability.

task-4350870

closes odoo/enterprise#84820

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>