[FIX] l10n_au_hr_payroll*: Fix size of Payroll IDs for Australian Payroll
The size of fields for Previous Payroll IDs was set to 20
characters. According to the STP xml schema, the max size is 200.
This is already check with the schema validation so the limit can
be safely removed.

This commit also adds a disclaimer for the users informing them
regarding STP compliance.

4820852

closes odoo/enterprise#88317

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Muhammad Haris Choudhary (mhch) <mhch@odoo.com>