[REV] account_disallowed_expenses: Revert some of the disallowed behavior.
Problem
---------
In odoo/enterprise#76066, we allowed to account for >100% deductibility,
but it wasn't done quite right. 

The whole purpose of the report and Fiscal categories is to help the
accountant to prepare his Fiscal Declaration.

In Belgium, those Fiscal Categories are used in that declaration to
declare:
- additional deductions (like the one for eInvoicing software)
- disallowed expenses (expenses that cannot be deduced from the
calculation of your income)
- ... detail your income to what's taxable and and at what rate (not yet
fully supported)

We considered that the added deductibility of 20% for invoicing software
expenses was a 120% Disallowed Expense, but this is wrong and that does
not exist.

It is just a 20% rate on particular software expense that can be
additionally reported in a specific section of the Fiscal Declaration,
and that will further reduce the income before the computation of the
income tax (the other 100% part being reported normally).

Objective
---------
Rename report columns to make the UI more generic and correct and
remove the disllowed computation that are not necessary anymore.

task-4848518

closes odoo/enterprise#88049

Related: odoo/upgrade#7879
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>