[FIX] account_winbooks_import : prevent problem on Winbooks import
Before this commit, there was a traceback error when importing Winbooks having a different sign on the balance in company currency and foreign currency.
There was a frozendict error in the import_wizard/_import_move().

Steps to reproduce:
- Choose a Belgium company
- Import a Winbooks file having a different sign on the balance in company currency and foreign currency
- Adapt account in Chart of Account if needed
- The traceback should be triggered

What happens is that the record where added to a list, then the app tries to change 'CURRAMOUNT' value to 0.
But it was a frozendict item, so it create an error.

I replace the record added to the list with a copy version that can be mutated.

opw-4794471

closes odoo/enterprise#88809

X-original-commit: e97aee9953b284ad82ee24bad29cd0922ba78314
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Corentin Heinix (cohe) <cohe@odoo.com>