[FIX] l10n_it_xml_export: Traceback on closing entry post
Description of the issue this commit addresses:

Since the new Tax Returns features in saas-18.3, a method that was used by the
method that was generating the xml export of the italian periodic vat report
was replaced by one of a different name. The one used by the xml export system
has not been modified and therefore tries to call a method that doesn't exist
leading to a traceback.

Steps to reproduce:

1. Install l10n_it_xml_export.
2. Using an Italian company, go to the montly vat report.
3. Create the closing entry and post it.
4. A traceback shows up.

Desired behavior after this commit is merged:

No traceback shows up during the xml export of the italian periodic vat
report closing entry posting process.

task-4825706

closes odoo/enterprise#87114

X-original-commit: 3a5335193b1cb0aa59b559c39a730b7d0f988c82
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>