[FIX] l10n_be_reports: 281.50 add missing countries
The aim of this commit is to provide a NIS code mapping for every
country created by default in Odoo.

Before this commit:
If a 281.50 was created for someone in one of these countries, it would
miss the `f2112_buitenlandspostnummer` field in the xml file resulting
in the administration rejecting the file.

After this commit:
A pragmatic completion is made for those countries in
`f2112_buitenlandspostnummer`.

closes odoo/enterprise#87267

Task: None
X-original-commit: 370acb83393f9c764b6a1f30cc426f282735535c
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>