[FIX] pos_enterprise: avoid error when preparation display has a single stage
This error occurs when we try to mark a single stage as ‘Reset’ or ‘Done’ in
the preparation display.

Steps to reproduce:
- Install the 'pos_restaurant' module
- Create a New 'Preparation Display' with one stage
- Open 'Preparation Screen'
- Now, 'Reset' or 'Done' the stage of order

Traceback:
IndexError: tuple index out of range

At [1], an error occurs because it tries to access a 'position' that doesn't
exist in the tuple. This happens because at [2], the code attempts to retrieve
the second-to-last (-2) stage position, but only one stage is being used.

[1]- https://github.com/odoo/enterprise/blob/285cca92a52f7b79de1d020558aa9b116cd7e44a/pos_enterprise/models/pos_prep_stage.py#L21-L22
[2]- https://github.com/odoo/enterprise/blob/285cca92a52f7b79de1d020558aa9b116cd7e44a/pos_enterprise/models/pos_prep_state.py#L72

sentry-6681171781

closes odoo/enterprise#92801

X-original-commit: 9d79999407861da5f3b9fc853ea935db2ffc5ef3
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>