[IMP] pos_iot: use websocket to print receipt
As for the blackbox, we now use the websocket as fallback to longpolling to print receipt.

This commit also fixes an issue regarding the module `pos_blackbox_be`: if the module was installed, but no blackbox
was configured, we could not print a receipt as no signature was provided.

closes odoo/enterprise#87240

Task: 4824066
X-original-commit: 4a195a014ccf63e434d1c6b8fcd20139e518cc32
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>