[FIX] l10n_de_reports: split vat into 2 columns if vat contains alphanumerics
Currently, the VAT is split into two columns only if the VAT code
consists solely of numeric values. However, in practice, VAT code
may contain a combination of letters and numbers.

Steps to reproduce :

1) Install l10n_de_reports ans switch to DE company
2) Create a partner with VAT as `NL000099998B57` from contacts
3) Now create a confirmed invoice for the above created partner
4) Go to "Accounting / Reporting / Audit Reports / General Ledger"
5) Download "DATEV DATA (ZIP)" and check "EXTF_customer_account.csv"

Issue:

Notice that the VAT was not split into country and VAT code for the above
created partner.

Cause:

Previously, the code only split the VAT into country code and VAT code
if the first two characters were alphabetic and the remainder was numeric.
This approach is insufficient because many countries have VAT codes that
include both letters and numbers.

Solution:

The code now uses the partner’s `_split_vat` method to separate the country code
and VAT code whenever a VAT is provided and has more than two characters.
It then validates the split VAT using the `simple_vat_check` helper method.

If the VAT is valid, the country code (in uppercase) and the VAT code are output
in separate columns; otherwise, the original VAT value is kept.

This change ensures that VAT numbers with alphanumeric codes are correctly
handled and split into their respective columns.

opw-4820929

closes odoo/enterprise#87563

X-original-commit: 3c8ada4e6b74a2ea3a1b3d4a5130f1e3c501dbed
Signed-off-by: John Laterre (jol) <jol@odoo.com>