[FIX] industry_fsm: allow resending single fsm task report
Currently, once a fsm task report has been sent once, the "Send report"
buttons are disabled in the form view. However, the contextual server
action is always available, but a check in `action_send_report` prevents
the report from being sent if the buttons are disabled.

This commits allows re-sending the report, which can be useful if some
changes have been made after the report was first sent.

opw-4818953

closes odoo/enterprise#88899

X-original-commit: bb6c19f5c833d134709689f9fbf3140f25dd53fe
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>