[FIX] l10n_be_reports: 281.50 pdf use debtor where required
The aim of this commit is to display the debtor info on the 281.50 pdf
report instead of the sender info.

Context:
If there is a representative for the company, he will be set as the
sender of the 281.50 and 325 instead of the company.

Before this commit:
The sender information was used in the pdf instead of the debtor
information.

After this commit:
The debtor information are used as required.

opw-4852895

X-original-commit: be56cfbcce4a2333c496f1e77ff244872e4c9572
Part-of: odoo/enterprise#87255
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>